Rename optuna.storage.rdb to optuna.storage._rdb.
