+++ optuna/storages/__init__.py
@@ -3,7 +3,7 @@ from typing import Union  # NOQA
 from optuna.storages._base import BaseStorage  # NOQA
 from optuna.storages._cached_storage import _CachedStorage
 from optuna.storages._in_memory import InMemoryStorage
-from optuna.storages.rdb.storage import RDBStorage
+from optuna.storages._rdb.storage import RDBStorage
 from optuna.storages._redis import RedisStorage
