+++ optuna/storages/_cached_storage.py
@@ -10,7 +10,7 @@ from typing import Tuple
 
 from optuna import distributions
 from optuna.storages import BaseStorage
-from optuna.storages.rdb.storage import RDBStorage
+from optuna.storages._rdb.storage import RDBStorage
 from optuna.study import StudyDirection
 from optuna.study import StudySummary
 from optuna.trial import FrozenTrial
