+++ optuna/storages/_rdb/storage.py
@@ -27,7 +27,7 @@ import optuna
 from optuna import distributions
 from optuna.storages._base import BaseStorage
 from optuna.storages._base import DEFAULT_STUDY_NAME_PREFIX
-from optuna.storages.rdb import models
+from optuna.storages._rdb import models
 from optuna.study import StudyDirection
 from optuna.study import StudySummary
 from optuna.trial import FrozenTrial
