+++ tests/storages_tests/rdb_tests/test_models.py
@@ -5,13 +5,13 @@ from sqlalchemy import create_engine
 from sqlalchemy.exc import IntegrityError
 from sqlalchemy.orm import Session
 
-from optuna.storages.rdb.models import BaseModel
-from optuna.storages.rdb.models import StudyModel
-from optuna.storages.rdb.models import StudySystemAttributeModel
-from optuna.storages.rdb.models import TrialModel
-from optuna.storages.rdb.models import TrialSystemAttributeModel
-from optuna.storages.rdb.models import TrialUserAttributeModel
-from optuna.storages.rdb.models import VersionInfoModel
+from optuna.storages._rdb.models import BaseModel
+from optuna.storages._rdb.models import StudyModel
+from optuna.storages._rdb.models import StudySystemAttributeModel
+from optuna.storages._rdb.models import TrialModel
+from optuna.storages._rdb.models import TrialSystemAttributeModel
+from optuna.storages._rdb.models import TrialUserAttributeModel
+from optuna.storages._rdb.models import VersionInfoModel
 from optuna.study import StudyDirection
 from optuna.trial import TrialState
